shape: rename from graph/shape to query/shape
