+++ query/shape/path.go
