+++ query/shape/shape.go
